requirements.txt
# Core
fastapi
uvicorn
requests

# Elastic + embeddings
sentence-transformers
pymupdf
# UI
streamlit

# Optional helpers
python-dotenv
gdown
python-dotenv
